requirements.txt
django-cors-headers==3.7.0
djangorestframework-simplejwt==5.0.0
asgiref==3.7.2
certifi==2024.6.2
charset-normalizer==3.3.2
cm==0.1a2
Django==4.2.16
djangorestframework==3.15.2
gunicorn==23.0.0
idna==3.7
packaging==24.1
pillow==10.3.0
psycopg2-binary==2.9.9
requests==2.32.3
sqlparse==0.4.4
typing_extensions==4.12.2
tzdata==2023.4
urllib3==2.2.2
whitenoise==6.7.0
